Call of Duty Gaming Data- Revision 1
The mission for this project was to become more aware of the range of data transformation and pre-processing that's available with PyCaret. After some great comments on my first draft from awesome people on LinkedIn, I was able to identify my errors. The next step should include more observations to bolster the training set and fluctuate pertinent in-game statistics (Kills, Deaths, and Duration).